tests/requirements.txt
pytest>=7.0
pytest-html>=4.0
requests>=2.28
python-dotenv>=1.0
selenium>=4.10
webdriver-manager>=4.0
